# OpenAPI.yml
openapi: '3.0.3'
info:
  title: 'Final Project Alta: Barengin'
  version: '1.0'
  description: 'API Documentation for Barengin'
  contact: 
    name: Barengin Official
    url: https://barengin.site
    email: sahrilmahendra@gmail.com
...
servers:
  - url: https://barengin.site
    description: Base URL that will be used for endpoints
...
paths:
...
  /products:
    get:
      tags:
      - Products
      summary: Get all products
      responses:
        '200':
          $ref: '#/components/responses/getAllProductsSuccess'
        '400':
          $ref: '#/components/responses/getAllDataFailed'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by id
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          $ref: '#/components/responses/getProductSuccess'
        '400':
          $ref: '#/components/responses/getDataByIdFailed'
...
  /jwt/products/{id}:
    put:
      tags:
        - Products
      summary: Update product by id for admin only
      security:
        - JWTAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        $ref: '#/components/requestBodies/requestBodyProductUpdate'
      responses:
        '200':
          $ref: '#/components/responses/successOperationNonData'
        '400':
          $ref: '#/components/responses/updateProductFailed'
        '500':
          $ref: '#/components/responses/productServerFailed'
    delete:
      tags:
      - Products
      summary: Delete product by id for admin only
      security:
        - JWTAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          $ref: '#/components/responses/successOperationNonData'
        '400':
          $ref: '#/components/responses/deleteDataByIdFailed'
####################################
#  GROUP PRODUCTS
####################################
...
  /products/group/products/{id_products}:
    get:
      tags:
        - Group Products
      summary: Get All Group Products by id products
      parameters:
        - $ref: '#/components/parameters/idProducts'
      responses:
        '200':
          $ref: '#/components/responses/getAllGroupProductsSuccess'
        '400':
          $ref: '#/components/responses/getDataByIdFailed'
...
components:
  schemas:
...
    requestDataProductUpdate:
      type: object
      required:
        - name_product
        - detail_product
        - price
        - limit
      properties:
        name_product:
          type: string
          minLength: 1
        detail_product:
          type: string
          minLength: 1
        price:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 2
        photo:
          type: string
          format: binary
          description: "allowed file extension: jpg, png, jpeg && size <= 1050000 byte"
...
    responseProductWithData:
      type: object
      properties:
        Code:
          type: integer
        Data:
          type: object
          properties:
            ID:
              type: integer
            Name_Product:
              type: string
            Detail_Product:
              type: string
            Price:
              type: integer
            Limit:
              type: integer
            Photo:
              type: string
              format: binary
            Url: 
              type: string
        Message:
          type: string      
    responseProductWithDataArray:
      type: object
      properties:
        Code:
          type: integer
        Data:
          type: array
          items:
            type: object
            properties:
              ID:
                type: integer
              Name_Product:
                type: string
              Detail_Product:
                type: string
              Price:
                type: integer
              Limit:
                type: integer
              Photo:
                type: string
                format: binary
              Url: 
                type: string
        Message:
          type: string
...
    responseGroupProductWithDataArray:
        type: object
        properties:
          Code:
            type: integer
          Data:
            type: array
            items:
              type: object
              properties:
                ID:
                  type: integer
                ProductsID:
                  type: integer
                NameGroupProduct:
                  type: string
                Limit:
                  type: integer
                CapacityGroupProduct:
                  type: integer
                Price:
                  type: integer
                AdminFee:
                  type: integer
                TotalPrice:
                  type: integer 
                DurationGroup:
                  type: string
                Name_Product: 
                  type: string
                Status: 
                  type: string
                Url:
                  type: string
                GetOrder: 
                  type: array
                  nullable: true
                  items:
                    type: object
                    properties:
                      OrderID:
                        type: integer
                      UsersID:
                        type: integer
                      GroupProductID:
                        type: integer
                      Name:
                        type: string
          Message:
            type: string  
...
    responseNonData: 
      type: object
      properties:
        Code:
          type: integer
        Message:
          type: string 
  parameters:
...
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        example: 1
    idProducts:
      name: id_products
      in: path
      required: true
      schema:
        type: integer
        example: 1
...
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "use Token"  
  requestBodies:
...
    requestBodyProductUpdate:
      required: true
      content:
        multipart/form-data:
          schema:
            $ref: '#/components/schemas/requestDataProductUpdate'
...
  responses:
...
    successOperationNonData:
      description: Success Response Non Data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseNonData'
          example:
            Code: 200
            Message: Success Operation
...
    getAllDataFailed:
      description: Bad Request Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseNonData'
          examples:
            Data Not Found:
              $ref: '#/components/examples/dataNotFound'
            Bad Request:
              $ref: '#/components/examples/badRequest'
...
    getDataByIdFailed:
      description: Bad Request Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseNonData'
          examples:
            Invalid Id:
              $ref: '#/components/examples/invalidId'
            Data Not Found:
              $ref: '#/components/examples/dataNotFound'
            Bad Request:
              $ref: '#/components/examples/badRequest'
...
    deleteDataByIdFailed:
      description: Bad Request Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseNonData'
          examples:
            Access Forbidden:
                $ref: '#/components/examples/accessForbidden'
            Invalid Id:
              $ref: '#/components/examples/invalidId'
            Data Not Found:
              $ref: '#/components/examples/dataNotFound'
...
    productServerFailed:
      description: Internal Server Error Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseNonData'
          examples:
            Can't Connect to Server:
              $ref: '#/components/examples/cantConnecttoServer'
            Failed to Upload File:
              $ref: '#/components/examples/failedtoUploadFile'
    getAllProductsSuccess:
      description: Success Response Data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseProductWithDataArray'
          example:
            Code: 200
            Data: [
              { 
                ID: 3,
                NameProduct: Youtube-Music,
                DetailProduct: Lorem,
                Price: 300000,
                Limit: 6,
                Photo: Youtube-Music-2021-12-13T10:06:45.jpg,
                Url: https://storage.googleapis.com/barengin-bucket/Youtube-Music-2021-12-13T10:06:45.jpg
              },
              { 
                ID: 4,
                NameProduct: Youtube Premium,
                DetailProduct: Berlangganan 1 bulan,
                Price: 200000,
                Limit: 4,
                Photo: Youtube-Premium-2021-12-14T09:34:47.png,
                Url: https://storage.googleapis.com/barengin-bucket/Youtube-Premium-2021-12-14T09:34:47.png
              }
            ]
            Message: Success Operation
    getProductSuccess:
      description: Success Response Data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseProductWithData'
          example:
            Code: 200
            Data: { 
              ID: 3,
                NameProduct: Youtube-Music,
                DetailProduct: Lorem,
                Price: 300000,
                Limit: 6,
                Photo: Youtube-Music-2021-12-13T10:06:45.jpg,
                Url: https://storage.googleapis.com/barengin-bucket/Youtube-Music-2021-12-13T10:06:45.jpg
            }
            Message: Success Operation
    updateProductFailed:
      description: Bad Request Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseNonData'
          examples:
            Invalid Id:
              $ref: '#/components/examples/invalidId'
            Data Not Found:
              $ref: '#/components/examples/dataNotFound'
            Access Forbidden:
              $ref: '#/components/examples/accessForbidden'
            Bad Request:
              $ref: '#/components/examples/badRequest'
            File Extension Not Allowed:
              $ref: '#/components/examples/fileExtensionNotAllowed'
            Illegal File:
              $ref: '#/components/examples/illegalFile'
            Size File Too Big:
              $ref: '#/components/examples/sizeFileTooBig'
...
    getAllGroupProductsSuccess:
      description: Success Response Data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/responseGroupProductWithDataArray'
          example:
            Code: 200
            Data: [
              { 
                ID: 1,
                ProducstID: 3, 
                NameGroupProduct: Youtube-Music-1,
                Limit: 6,
                CapacityGroupProduct: 2,
                Price: 300000,
                AdminFee: 5000,
                TotalPrice: 305000,
                DurationGroup: 12-26-2021,
                Name_Product: Youtube-Music,
                Status: Available,
                Url: https://storage.googleapis.com/barengin-bucket/Youtube-Music-2021-12-13T10:06:45.jpg,
                GetOrder: [
                  {
                    OrderID: 1,
                    UsersID: 79,
                    GroupProductID: 1,
                    Name: nuril
                  }
                ]
              },
              { 
                ID: 2,
                ProducstID: 3, 
                NameGroupProduct: Youtube-Music-2,
                Limit: 6,
                CapacityGroupProduct: 1,
                Price: 300000,
                AdminFee: 5000,
                TotalPrice: 305000,
                DurationGroup: 12-26-2021,
                Name_Product: Youtube-Music,
                Status: Available,
                Url: https://storage.googleapis.com/barengin-bucket/Youtube-Music-2021-12-13T10:06:45.jpg,
                GetOrder: null
              }
            ]
            Message: Success Operation
...
  examples:
...
    accessForbidden:
      value:
        Code: 400
        Message: Access Forbidden                
    badRequest:
      value:
        Code: 400
        Message: Bad Request
    cantConnecttoServer:
      value:
        Code: 500
        Message: Can't Connect to server
...
    dataNotFound:
      value:
        Code: 400
        Message: Data Not Found
...
    failedtoUploadFile:
      value:
        Code: 500
        Message: Failed to Upload File
    fileExtensionNotAllowed:
      value:
        Code: 400
        Message: File Extension Not Allowed
...
    illegalFile:
      value:
        Code: 400
        Message: Illegal File
...
    invalidId:
      value:
        Code: 400
        Message: Invalid Id
...
    sizeFileTooBig:
      value:
        Code: 400
        Message: Size File Too Big